Add support check on cleanup
Before, the event listener would not be added if the element did not support adding event listeners, but it would error on cleanup when it tries to remove the event listener (`removeEventListener` is not a function). This adds the same check to the cleanup and makes sure that it bails out early if event listeners are not supported.